refactor(path): seperate index & Page
